consolidated ERC-20 code into single folder
